requirements.txt
discord.py==1.3.4
PyNaCl==1.4.0
python-dotenv==0.14.0
firebase-admin==2.18.0
marshmallow==3.7.0
marshmallow-dataclass==7.6.0
APScheduler==3.6.3
pytz==2020.1
